util function added
